Factor testing facade interaction out of ControllerTest
(cherry picked from commit 252c8e338fb48b5331a8aa8daa8f9636e48dfd37)
